backend/requirements.txt
Flask
openai
supabase
realtime
python-dotenv